fix prompt_toolkit attribute tab completion
The prompt_toolkit_completer was using an incorrect begin_index for
matching completions except when the command typed was the first on a
given line.

Now it will complete subcommands, provided they are available in
`$BASH_COMPLETIONS` -- that is, if you type

```
git s
```

and hit TAB, you should see `status`, `shell`, `stash`, etc. given as
options.